Replace encoding and hashing modules with ones that use bit32 and remove Lockbox
Roblox has gotten a port of the bit32 library from Lua 5.2, rendering the Lua implemention of a bitwise library unnecessary and slow.

Given the existing lockbox code had several problems (namely that it was written to be readable, not fast; and that the license was not respected), I decided to simply rewrite the modules from it instead of modifying them to use bit32.